Moves maturity content to its own structure
