+++ src/components/EntityOpsLevelMaturityContent/ServiceMaturityReport/Campaigns/CampaignSummary.tsx
@@ -7,7 +7,7 @@ import CheckCircleIcon from "@material-ui/icons/CheckCircle";
 import CancelIcon from "@material-ui/icons/Cancel";
 import { useApi, configApiRef } from "@backstage/core-plugin-api";
 import { TeamOutlined } from "@ant-design/icons";
-import { ChecksByCampaign } from "../types/OpsLevelData";
+import { ChecksByCampaign } from "../../../../types/OpsLevelData";
 import ColoredCampaignStatus from "./ColoredCampaignStatus";
 
 interface CampaignSummaryProps {
