+++ src/components/EntityOpsLevelMaturityContent/ServiceMaturityReport/CheckResultsByLevel/CheckResultDetails.test.tsx
@@ -7,7 +7,7 @@ import {
 } from "@backstage/test-utils";
 import { configApiRef } from "@backstage/core-plugin-api";
 import CheckResultDetails from "./CheckResultDetails";
-import { CheckResult, CheckResultStatus } from "../types/OpsLevelData";
+import { CheckResult, CheckResultStatus } from "../../../../types/OpsLevelData";
 
 const mockConfig = new MockConfigApi({
   opslevel: { baseUrl: "https://example.com" },
