+++ src/components/EntityOpsLevelMaturityContent/ServiceMaturityReport/CheckResultsByLevel/CheckResultsByLevel.test.tsx
@@ -7,7 +7,7 @@ import {
 } from "@backstage/test-utils";
 import { configApiRef } from "@backstage/core-plugin-api";
 import CheckResultsByLevel from "./CheckResultsByLevel";
-import { LevelCheckResults } from "../types/OpsLevelData";
+import { LevelCheckResults } from "../../../../types/OpsLevelData";
 
 const mockConfig = new MockConfigApi({
   opslevel: { baseUrl: "https://example.com" },
