+++ src/components/EntityOpsLevelMaturityContent/ServiceMaturityReport/EntityOpsLevelMaturityProgress.tsx
